refactor(engine): wrap code piece execution in a sandbox
