+++ packages/engine/src/lib/handler/code/code-executor-common.ts
@@ -1,0 +1,12 @@
+export type CodeModule = {
+    code(input: unknown): Promise<unknown>
+}
+
+export type CodeExecutorSandbox = {
+    run(params: RunParams): Promise<unknown>
+}
+
+type RunParams = {
+    codeModule: CodeModule
+    input: Record<string, unknown>
+}
