+++ packages/engine/src/lib/handler/code/code-executor-sandbox.ts
@@ -1,0 +1,4 @@
+import { CodeExecutorSandbox } from './code-executor-common'
+import { noOpCodeExecutorSandbox } from './no-op-code-executor-sandbox'
+
+export const codeExecutorSandbox: CodeExecutorSandbox = noOpCodeExecutorSandbox
