+++ packages/engine/src/lib/handler/code/code-executor.ts
@@ -1,11 +1,9 @@
 import { ActionType, CodeAction, GenericStepOutput, StepOutputStatus } from '@activepieces/shared'
-import { BaseExecutor } from './base-executor'
-import { ExecutionVerdict, FlowExecutorContext } from './context/flow-execution-context'
-import { EngineConstants } from './context/engine-constants'
-
-type CodePieceModule = {
-    code(params: unknown): Promise<unknown>
-}
+import { BaseExecutor } from '../base-executor'
+import { ExecutionVerdict, FlowExecutorContext } from '../context/flow-execution-context'
+import { EngineConstants } from '../context/engine-constants'
+import { CodeModule } from './code-executor-common'
+import { codeExecutorSandbox } from './code-executor-sandbox'
 
 export const codeExecutor: BaseExecutor<CodeAction> = {
     async handle({
@@ -20,19 +18,27 @@ export const codeExecutor: BaseExecutor<CodeAction> = {
         if (executionState.isCompleted({ stepName: action.name })) {
             return executionState
         }
-        const { censoredInput, resolvedInput } = await constants.variableService.resolve({
+
+        const { censoredInput, resolvedInput } = await constants.variableService.resolve<Record<string, unknown>>({
             unresolvedInput: action.settings.input,
             executionState,
         })
+
         const stepOutput = GenericStepOutput.create({
             input: censoredInput,
             type: ActionType.CODE,
             status: StepOutputStatus.SUCCEEDED,
         })
+
         try {
             const artifactPath = `${constants.baseCodeDirectory}/${action.name}/index.js`
-            const codePieceModule: CodePieceModule = await import(artifactPath)
-            const output = await codePieceModule.code(resolvedInput)
+            const codeModule: CodeModule = await import(artifactPath)
+
+            const output = await codeExecutorSandbox.run({
+                codeModule,
+                input: resolvedInput,
+            })
+
             return executionState.upsertStep(action.name, stepOutput.setOutput(output)).increaseTask()
         }
         catch (e) {
