+++ packages/engine/src/lib/handler/code/no-op-code-executor-sandbox.ts
@@ -1,0 +1,10 @@
+import { CodeExecutorSandbox } from './code-executor-common'
+
+/**
+ * Runs code without a sandbox.
+ */
+export const noOpCodeExecutorSandbox: CodeExecutorSandbox = {
+    run({ codeModule, input }) {
+        return codeModule.code(input)
+    },
+}
