+++ packages/engine/src/lib/handler/flow-executor.ts
@@ -1,5 +1,5 @@
 import { Action, ActionType, isNil } from '@activepieces/shared'
-import { codeExecutor } from './code-executor'
+import { codeExecutor } from './code/code-executor'
 import { ExecutionVerdict, FlowExecutorContext } from './context/flow-execution-context'
 import { branchExecutor } from './branch-executor'
 import { BaseExecutor } from './base-executor'
